requirements.txt
coverage==5.1
django-crispy-forms==1.9.1
django-test-plus==1.4.0
django==3.0.7
factory-boy==2.12.0
model-bakery==1.1.0
pytest-cov==2.9.0
pytest-django==3.9.0
pytest==5.4.3
selenium==3.141.0
splinter==0.13.0